% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dconf.R
\name{dconf}
\alias{dconf}
\title{Find G-Coefficient Confidence Interval}
\usage{
dconf(data, col.scores, n, conf.level = 0.95, rounded = 3)
}
\arguments{
\item{data}{A data frame containing a column for "Person", a column for "Trial", and the remaining columns for metrics.}

\item{col.scores}{A column index or variable name denoting the column in the data frame that serves as the dependent variable.}

\item{n}{The number of trials to test.}

\item{conf.level}{Confidence level (in decimal form).}

\item{rounded}{The number of decimal places the reliability coefficients should be rounded to.}
}
\value{
A data frame containing the G-coefficient (for absolute agreement), as well as the corresponding lower and upper bounds at the specified confidence level.
}
\description{
Find G-Coefficient Confidence Interval
}
\note{
The confidence interval calculation, originally developed by Arteaga et al. (1982) assumes that score effects are normally distributed. Hence, any attempt to use dconf() with non-normal data will not produce a reasonable result!
}
\examples{
x <- data.frame(c(1,1,1,2,2,2,3,3,3), c(1,2,3,1,2,3,1,2,3), c(12, 14, 12, 22, 22, 19, 17, 22, 20))
colnames(x) <- c("Person", "Trial", "Metric")
dconf(x, col.scores = "Metric", n = 5, conf.level = .9)
}
