% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstudy.R
\name{dstudy}
\alias{dstudy}
\title{Execute a D-Study}
\usage{
dstudy(data, col.scores, from, to, by, rounded = 3)
}
\arguments{
\item{data}{A data frame containing a column for "Person", a column for "Trial", and the remaining columns for metrics.}

\item{col.scores}{A column index or variable name denoting the column in the original data frame that serves as the dependent variable.}

\item{from}{The first number of trials to test.}

\item{to}{The final number of trials to test.}

\item{by}{The interval at which trial numbers should be tested.}

\item{rounded}{The number of decimal places the reliability coefficients should be rounded to.}
}
\value{
A data frame of G-coefficients (absolute agreement) for the specified numbers of trials.
}
\description{
Execute a D-Study
}
\examples{
x <- data.frame(c(1,1,1,2,2,2,3,3,3), c(1,2,3,1,2,3,1,2,3), c(12, 14, 12, 22, 22, 19, 17, 22, 5))
colnames(x) <- c("Person", "Trial", "Metric")
dstudy(x, col.scores = "Metric", from = 1, to = 10, by = 1)
}
